top: support for multiple events and fields
Associate the event to the process and display it in the top mode, and
you can customize multiple events and their fields to be observed.

Processes have many parameters that are not exported in the proc/<pid>/
directory. We can use the fields exported by the kernel tracepoint, as
well as dynamically added fields via kprobe, to observe this information
of the process in the top mode.

The event attribute `top-add' adds a top column, and `top-by' adds a
sorted top column. Assign the field name of the tracepoint.

Run cat command to see all fields.
$ cat /sys/kernel/debug/tracing/events/sched/sched_stat_wait/format

Example
$ sysctl kernel.sched_schedstats=1  #enable schedstats
$ perf-prof top -e "sched:sched_stat_iowait//top-add=delay/,sched:sched_stat_sleep//top-add=delay/,sched:sched_stat_wait//top-by=delay/"
perf-prof - 00:36:46  sample 1989797 events
     PID        DELAY        DELAY        DELAY COMM
  127466            0   2545278377    158442485 pal_session
  141483            0   2880332128    104786645 python

PID  DELAY  DELAY  DELAY  COMM
     |      |      `sched:sched_stat_wait delay field, time the task is runnable
     |      |       but not actually running due to scheduler contention
     |      `sched:sched_stat_sleep delay field, time the task is not runnable
     |       including iowait
     `sched:sched_stat_iowait delay field, time the task is not runnable due to
      waiting on IO to complete

Signed-off-by: duanery <corcpp@foxmail.com>
Signed-off-by: yongduan <yongduan@tencent.com>