new profiler order: order events by timestamp
Order is a decorator for other profilers. The sampled events are sorted by timestamp
and then provided to the underlying profiler. Can control the memory used by sorting.

In order to be able to automatically obtain the timestamp of the sampled event, the
perf_event_attr.sample_type of the underlying profiler must have the PERF_SAMPLE_TID
and PERF_SAMPLE_TIME bits set.

Example
$ perf-prof trace -e sched:sched_wakeup --order --order-mem=1

Signed-off-by: duanery <corcpp@foxmail.com>