kmemleak: support LEAKED BYTES REPORT
The size=bytes_alloc attribute is used to get the allocated memory size.
Leaked bytes is all allocated memory that is not freed.

Using keyvalue_pairs, the key is the memory allocation stack of the event,
and the value is the leaked bytes corresponding to the stack.

Finally, sort by the leaked bytes, and output the report from largest leak
to smallest.

Example:
$ perf-prof kmemleak --alloc "kmem:kmalloc,kmem:kmalloc_node" --free kmem:kfree \
  -m 64 -g --order --order-mem 4M
KMEMLEAK STATS:
TOTAL alloc 715 free 834

LEAKED BYTES REPORT:
Leak of 2880 bytes in 45 objects allocated from:
    ffffffff811726db __kmalloc+0x20b ([kernel.kallsyms])
    ffffffff81232bc4 ext4_htree_store_dirent+0x34 ([kernel.kallsyms])
    ffffffff81242a69 htree_dirblock_to_tree+0x169 ([kernel.kallsyms])
    ffffffff81243b92 ext4_htree_fill_tree+0xe2 ([kernel.kallsyms])
    ffffffff81232515 ext4_readdir+0x185 ([kernel.kallsyms])
    ffffffff8119f880 vfs_readdir+0x80 ([kernel.kallsyms])
    ffffffff8119fcf5 sys_getdents+0x95 ([kernel.kallsyms])
    ffffffff81ad6db9 system_call_fastpath+0x16 ([kernel.kallsyms])

Note: ext4_htree_store_dirent is not a real memory leak, just an example.

Signed-off-by: duanery <corcpp@foxmail.com>
Signed-off-by: yongduan <yongduan@tencent.com>