multi-trace: Extend --detail to display events before event1
--detail=-1ms, Print events within 1ms before event1.

With the --order feature, on the timeline, the backup events are also
added to a doubly linked list. Events in the doubly linked list are
ordered by time. Therefore, 1ms before the first element of the list
is unneeded and can be safely released.

When paired events are output to two_event_class::two, in addition to
events between event1 and event2, events within 1ms before event1 are
also included. It is up to the impl to decide how to handle the event.

- delay impl, above the threshold(--than), print all events.

Signed-off-by: duanery <corcpp@foxmail.com>
Signed-off-by: yongduan <yongduan@tencent.com>