profiler: hrcount: high resolution counter
Start the hrtimer through cpu-clock, and sample the value of the
counter regularly to obtain high-resolution counts.

In user mode, the count period is specified by the --period parameter,
and the difference between the previous count and the next count is
calculated and displayed.

Example:
$ perf-prof hrcount -e sched:sched_wakeup,sched:sched_switch,kvm:kvm_exit,\
  kvm:kvm_entry -C 1 --period 100ms -i 1000
2022-09-04 22:57:59.744242
sched:sched_wakeup   14|  18|  39|  16|  10|  26|  26|  18|  18|   8
sched:sched_switch   46|  64|  90| 116| 166|  62|  74|  52|  42|  18
      kvm:kvm_exit 1539|1469|1641|1474|1445|1478|1288|1285|1445|1434
     kvm:kvm_entry 1539|1469|1641|1473|1446|1478|1288|1285|1445|1434
2022-09-04 22:58:00.744236
sched:sched_wakeup   46|  26|  22|  39|   9|  17|   9|  39|  98|  47
sched:sched_switch  115|  58|  72| 134|  50|  34|  58| 106| 248| 114
      kvm:kvm_exit 1444|1386|1479|1705|1652|1527|1366|1514|1967|1551
     kvm:kvm_entry 1444|1386|1479|1705|1652|1527|1366|1514|1967|1551

Counting with a period of 100m, print 10 count values per second.

With a shorter period, observations can be made at the microscopic scale.
For example, microbursts.

Signed-off-by: duanery <corcpp@foxmail.com>
Signed-off-by: yongduan <yongduan@tencent.com>