lib/perf: to support perf_event cgroup
When the PERF_FLAG_PID_CGROUP flag is set, the perf_event_open syscall
uses the pid as a cgroup fd.  Therefore, you can use the pid of
perf_thread_map as a cgroup fd and set PERF_FLAG_PID_CGROUP flag.

The perf_event cgroup was originally designed to select a group of pids.

Signed-off-by: duanery <corcpp@foxmail.com>