oncpu: Use sched:sched_switch to calculate the runtime of the process.
       ps   1214 d... [000]  2359.771892: sched:sched_switch: ps:1214 [120] R ==> sap1001:112746 [120]
  sap1001 112746 d... [000]  2359.772143: sched:sched_switch: sap1001:112746 [120] S ==> ps:1214 [120]
       ps   1214 d... [000]  2359.772335: sched:sched_switch: ps:1214 [120] R ==> tbsd_mcd:54713 [120]

The runtime of sap1001:112746 on CPU000 is equal to 2359.772143 minus 2359.771892.
And so on, the runtime of ps:1214 can also be calculated.

Therefore, you can only count the processes that have run on the specified CPU, and
the runtime of the process. Arrange from largest to smallest.

E.g.

CPU SUM(ms) COMM:TID(ms)
000 23      ps:7155(19ms) check_dfwagent.:7147(1ms) grep:7150(0ms) sap1009:112762(0ms) python:58876(0ms)
024 7       pal_session:55606(2ms) pal_ctrl:55607(0ms) check_dfwagent.:7158(0ms) AdamPlugins:7147(0ms)

Signed-off-by: duanery <corcpp@foxmail.com>