watchdog: Cut the CPU list according to watchdog_cpumask.
Commit 93f770a (Pass cpus and threads to monitor) create perf_cpu_map before
monitor->init. Therefore, inside watchdog.init, reading 'watchdog_cpumask'
to env->cpumask cannot take effect immediately. As a result, the watchdog
always stays in the INIT stage.

This commit fixes the problem, watchdog_cpumask is the list of CPUs that
actually enable watchdog, and only a subset can be selected. Therefore,
take the intersection with the -C parameter, and then convert it back to
a string, and assign it to env->cpumask.

Fixes: 93f770a (Pass cpus and threads to monitor)
Signed-off-by: duanery <corcpp@foxmail.com>