multi-trace: ATTR trigger, use events to trigger interval output.
First remove the '-i' parameter. Then, add the trigger attribute to the
periodic event to periodically trigger the interval output. It can also
be added to a special event as a boundary, to print previously processed
events.

As a trigger event, the processed information can be printed at a very
precise point in time.

Signed-off-by: duanery <corcpp@foxmail.com>