multi-trace: Use a separate two(A,NULL) to carry the remaining events.
Previously, -e A -e B,C would create two(A,B), two(A,C) two objects.
Randomly choose (A,B) or (A,C) to carry the remaining events.

Now, two(A, NULL) objects will be created to carry the remaining
events, which is more reasonable.

Currently, only pair and kmemprof need to carry the remaining events.
pair directly prints the remaining events. kmemprof, the remaining
events are allocated but not released, which can be regarded as a
memory leak.

You can also use the only-remaining mode, -e A -e B//untraced/, only
create a two (A, NULL) object. E.g:

  perf-prof multi-trace -e sched:sched_switch -e kprobes:calc_global_load \
  //untraced/trigger/ -m 256 --order --order-mem 32M --impl pair

  Only print the last sched_switch event that occurred on the CPU. From
  this you can get the process running on the current cpu.

Signed-off-by: duanery <corcpp@foxmail.com>