hrtimer: Use {expression} to decide whether to print samples.
High-resolution timer sampling. During the sampling interval, it is up to
whether the events occurs or not to print samples. Whether the event occurs
or not is determined by {expression}. The expression uses the event name as
a variable, which represents the number of occurrences within the specified
period. If the expression is true, the sample is printed, otherwise it is
not printed.

Example
  perf-prof hrtimer -e sched:sched_switch -C 0 --period 50ms 'sched_switch==0'

Signed-off-by: duanery <corcpp@foxmail.com>