lib/epoll: When freeing data, check whether data is being used in ep->events.
Inside event_poll__poll(), events A, B, C are being handled sequentially. If
inside the `handle` of event A, execute event_poll__del() to delete event B.
At this point, the `data` object of event B will be released. When
event_poll__poll() goes to handle event B, perf-prof will crash because of
use-after-free.

When deleting the data object of event B, check whether the data is in use
in ep->events, and if so, clear it.

Signed-off-by: duanery <corcpp@foxmail.com>