monitor: Support net-events through push and pull attributes.
push attribute. Start a tcp service, and broadcast the perf event of the tp
through tp_broadcast_event().

pull attribute. Pull events from the tcp server, convert the events, and
finally pass them to the profiler. 'stream_id' and 'common_type' are two
fields that must be converted. These two fields are only meaningful on the
server side, and are replaced by the local values of tp. When disconnected
from the tcp server, re-enable kernel events.

In a virtualization scenario, push and pull are generally used to combine
guest events and host events to cooperate with profile analysis.

Signed-off-by: duanery <corcpp@foxmail.com>