monitor: vm attr, get the mapping from Guest vcpu to Host tid.
A new multi-purpose vm attribute is added to mark which client the event
comes from, and secondly, it is used to convert Guest vcpu to Host thread
id.

This vm marker can be printed when printing events.

The Guest vcpu to Host tid is used to filter events. Events that occur on
the specified vcpu in the Guest can be regarded as events occurring on the
thread corresponding to the Host. Host can filter the thread id to filter
events on the specified Guest vcpu.

Signed-off-by: duanery <corcpp@foxmail.com>