multi-trace: samekey: Only show events with the same key as event1 or event2.
If the events come from the kernel, they can be compared directly. If the
event comes from Guest, you need to add the vm attr to convert the event,
and then it can be compared.

Samekey feature, compare the key of events. For untraced events, you also
need to add the key attr. For Guest events, you also need to add the vm attr
to convert the event. If the key represents a thread id, you can filter out
Guest events that occur on the specified vcpu.

For example:
sched:sched_switch//untraced/pull=9900/vm=vm_uuid/key=common_pid/ --detail=samekey

For the sched:sched_switch events from Guest 'vm_uuid', its 'common_pid' will
be converted to Host tid, and then the key attribute will be used to obtain
the tid and compare it with the key of event1 and event2.

Signed-off-by: duanery <corcpp@foxmail.com>