multi-trace: Know in advance if unpaired events are delayed more than a specified time.
Recording the timestamp of the latest event can be compared with the timestamp of the
`event1' backed up in `ctx.backup'. And print the events before and after event1.

Being able to predict delays in advance allows for more information to be obtained in
advance. Therefore, this is a very meaningful exploration.

Signed-off-by: duanery <corcpp@foxmail.com>