multi-trace: When --order is enabled, get the real event lost time.
When --order is enabled, events are queued and flushing is delayed by one round.
The lost event is not queued and is passed directly to the profiler. Therefore,
the tl_last->time obtained by timeline_free_unneeded() is not the correct event
loss time.

`ctx.recent_lost_time' stores the real event loss time. Any event before
`ctx.recent_lost_time' may be unpaired. In the next interval, if event2 is not
received, the event is considered lost and removed from the timeline forcibly.

Signed-off-by: duanery <corcpp@foxmail.com>