multi-trace: Start sampling after the events is fully enabled.
  -e sched:sched_wakeup -e sched:sched_switch -C 0-95

A sched_wakeup occurs on CPU0, possibly a paired sched_switch occurs on CPU95. When enabling,
CPU0 is enabled first, and CPU95 is enabled last. It is possible that the sched_wakeup event
is only sampled on CPU0, and the sched_switch event is not sampled on CPU95.

It is possible that sched_wakeup will block the timeline to free unneeded events.

Signed-off-by: duanery <corcpp@foxmail.com>