monitor: use stdin to trigger a marker and interval.
Markers are used to mark things that happen in the outside world into the
profiling process. Implemented, reads a line from stdin and outputs to stdout.
The output will be merged with the output of perf-prof.

The interval is like a barrier, isolating the profiling results before and
after the marker.

Signed-off-by: duanery <corcpp@foxmail.com>