Event supports new form: profiler[/option/ATTR/ATTR/.../]
Think of the profiler as a new event source and can be opened like a tracepoint.

Tracepoint is a kernel event source and is compatible with different kernel versions and
different platforms, and is relatively simple to use.

Hardware pmu event sources vary widely. Different platforms are not unified, have
different register names, and different hardware facilities. Profiler can package these
differences and support cross-platform, virtualization, and more hardware capabilities.
From this, the profiler can be viewed as a consistent event source. Opened and processed
like tracepoint.

The events sampled by the profiler are first converted into a unified structure, which
is "struct perf_record_dev". Then it is forwarded to the target prof_dev and processed
together with the tracepoint event, such as sorting, delayed analysis(multi-trace), etc.

This patch pre-modifies the trace profiler. Here is an example.

    perf-prof trace -e 'task-state,page-faults,raw_syscalls:sys_enter,profile/-F 1000/' \
    --order -- cat /proc/self/maps

    "raw_syscalls:sys_enter" is a tracepoint. Task-state, page-faults, and profile are
    three profilers. This command tracks the task switching, page faults, and system
    calls of the cat process while sampling at a frequency of 1000.

Add the --watermark <0-100> option to reduce the profiler's wake-up frequency and reduce
CPU utilization.

In short, profilers can be used independently or in combination to solve complex problems.

Signed-off-by: duanery <corcpp@foxmail.com>