The source device supports in time flush.
When the forwarding target device enables order, the events returned to the source
device are also ordered. Then, the source device does not need to enable order.

However, if the source device automatically enables order, such as profiler::order=1,
it needs to be flushed. Ensure that the source device can handle events in timely.

Signed-off-by: duanery <corcpp@foxmail.com>