monitor: Forwarding resets the instance when nr_ins does not match.
Like this command:
  perf-prof multi-trace -e XX:YYY -e XX:ZZZ,task-state//untraced/ -p 1234 --order \
            --than 10ms --detail=sametid

multi-trace is attached to pid 1234, task-state is also attached to pid 1234, but
internally it switches to the cpu.

In this scenario, forwarding will reset the instance. Task-state events are forwarded
to multi-trace, but the instance seen is 0.

The target device has to decide whether it can receive the forwarded event, and
instance=0.

Signed-off-by: duanery <corcpp@foxmail.com>