event-spread: prof_dev is flushed immediately when a new client connects.
$ perf-prof trace -e irq:softirq_entry//push=9900/ &
$ perf-prof multi-trace -e irq:softirq_entry//pull=9900/ -e irq:softirq_exit -i 1000 --order

push=9900, the event is sent to tcp port 9900, but without -i 1000, the ringbuffer will
not be flushed periodically.

pull=9900, get events from tcp 9900 port, but flush the ringbuffer in 1000ms. When order
is enabled, an "Out of order" warning will appear. This is caused by the fact that some
events from tcp port 9900 are flushed after the ringbuffer has been partially used.

Therefore, prof_dev is flushed immediately when a new client connects.

Signed-off-by: duanery <corcpp@foxmail.com>