multi-trace: use profiler//untraced/, order must be enabled.
profiler//untraced/ uses an independent ringbuffer, and when its events are
forwarded to multi-trace, the results will be incorrect.

For example:
  perf-prof multi-trace -e kvm:kvm_exit -e 'kvm:kvm_entry,task-state/-m 256/untraced/' \
  -t 210673 -m 128 -i 1000 --than 80us --detail=sametid

Multi-trace tracks the delay of kvm_exit => kvm_entry and displays task-state
events that occur in the middle of the delay. But only thread 210673 is tracked.

If order is not enabled, when multi-trace processes events, kvm:kvm_exit and
kvm:kvm_entry occur inside the ringbuffer corresponding to thread 210673. After
processing, the event is quickly removed from the timeline. When processing
events forwarded by task-state, there are no kvm events in the timeline. Leading
to incorrect results, such as: unable to detect the task-state event in the
middle of kvm_exit => kvm_entry.

This is due to profiler//unraced/ using a separate ringbuffer. Order must be
enabled, sort first and then process.

Signed-off-by: duanery <corcpp@foxmail.com>