multi-trace: Correctness is guaranteed when events are lost.
Currently, event loss only ensures that the timeline is not blocked, but
does not guarantee the correctness of the results. But in most cases,
events will be lost, and correctness is very necessary.

A basic principle: events sampled within the start and end time of lost are
unsafe. Events before and after this are safe.

           lost
  - - - -|= = = =|- - - -
     |   |  |    ` end_time, the first event after lost.
     |   |  ` lost, events in the lost range from different instances.
     |   ` start_time, the last unlost event.
     ` events before lost. It will be added to `ctx->backup' and block the
       timeline.

When lost, the events in ctx->backup are unsafe. For two(A, B), A is in
ctx->backup, and if B is > start_time, then two(A, B) is unsafe. B cannot
determine whether it happened again after it was lost.

Therefore, nothing can be done for events in the lost range, neither
`_find_prev` nor `_backup'.

Secondly, the events that have been backed up in ctx->backup must be deleted.
It can not only ensure correctness, but also avoid blocking the timeline.

See the comments of multi_trace_event_lost() for details.

Signed-off-by: duanery <corcpp@foxmail.com>