Atomic enable avoids enablement loss.
  -e sched:sched_wakeup -e sched:sched_switch -C 0-95

A sched_wakeup occurs on CPU0, possibly sched_switch occurs on CPU95. When
enabling, CPU0 is enabled first, and CPU95 is enabled last. It is possible
that the sched_wakeup event is only sampled on CPU0, and the sched_switch
event is not sampled on CPU95.

This will cause problems with profilers such as multi-trace and task-state.
They need to be enabled atomically, which means all CPUs open the event at
the same time. Likewise, atomic disabling is also required.

perf_evlist__enable() enables all cpu and thread events for each evsel and
cannot be atomic. Therefore, events that occur during the enabling need to
be discarded. It also means that only events after a certain moment need to
be processed.

This moment is an atomic point in time. Events before and after are atomically
separated. In fact, most profilers require such atomicity.

Atomic disabling only affects the exit of the perf-prof command and is not
supported at the moment.

Signed-off-by: duanery <corcpp@foxmail.com>