lib/perf: Make sure to follow the group event on the same CPU.
For all evsels, perf_evlist__set_maps() sets the same cpus and threads, and
perf_evlist__set_leader() sets the same leader. Therefore, the cpus of leader
and evsel are the same.

However, perf_evsel__set_own_cpus() can set a different cpus for evsel. But
it should be ensured that evsel->cpus is a subset of leader->cpus. The wrong
group_fd will be obtained in perf_evsel__open(), causing perf_event_open to fail.

CPU   leader(0-24)    evsel(0-24)    evsel(1-2)
0     leader.fd.0 --- evsel.fd.0 ---                 in a group, same ringbuffer.
1     leader.fd.1 --- evsel.fd.1 --- evsel.fd.0  --- in a group, same ringbuffer.
2     leader.fd.2 --- evsel.fd.2 --- evsel.fd.1  --- in a group, same ringbuffer.
3     leader.fd.3 --- evsel.fd.3              ` => idx = 1, leader idx = 2.
...   leader.fd..     evsel.fd..

This fix is used to convert the index of evsel->cpus into the index of leader->cpus,
and then obtain the group_fd correctly.

perf_evlist__mmap() can also support evsel setting different cpus, as follows:
	mmap_per_evsel() {
		cpu = perf_cpu_map__idx(evsel->cpus, evlist_cpu);
		if (cpu == -1)
		continue;
	}

So this is all that's needed.

Signed-off-by: duanery <corcpp@foxmail.com>