prof_dev: Wall-clock conversion of events.
First, the timestamp of event sampling is local_clock() and cannot be directly
converted to a wall-clock.

Secondly, the sampled events will be cached inside the ringbuffer or order mem.
Event processing will be delayed. When it is processed, gettimeofday() obtains
the wall-clock, not the event occurrence time.

Third, perf_event_attr::clockid = CLOCK_REALTIME will return -EINVAL for
tracepoint. This also fails to get the wall-clock of events.

We need a relatively correct event wall-clock, but don't require it to be
completely accurate. Our approach is to obtain the local_clock() and wall-clock
of a base event, and the wall-clock of the new events can be calculated.

    new_wall_clock = base_wall_clock + (new_local_clock - base_local_clock)

The deviation between the local_clock and the wall-clock of the base event is
fixed, and the wall-clock of the new event is relatively accurate.

There are 2 ways to obtain local_clock and wall-clock.

    1) TSC conversion. Get the wall-clock and tsc value at the same time, and
    convert tsc to local_clock.

    2) Get the wall-clock and immediately call the getpid() to get the local_clock
    for the syscalls:sys_enter_getpid event.

The new prof_dev_print_time() is used to convert the event local_clock to
wall-clock and print it.

Signed-off-by: duanery <corcpp@foxmail.com>