vm attr: Guest vcpu switches to Host cpu
  sched:sched_switch//pull=xx/vm=uuid/  # Pull events from Guest

The cpu of these events is switched to the host cpu, and the pid/tid is switched to
the pid/tid of the Host Hypervisor.

At the same time, obtain instances of these events. The obtained instance is the
instance corresponding to the profiler executed on the Host. Discard events for which
the instance cannot be obtained. Mainly used in multi-trace profilers.

For example.

  Guest:
  perf-prof trace -e 'sched:sched_switch//push="/dev/virtio-ports/org.qemu.perf0"/' --tsc

  Host:
  perf-prof multi-trace -e sched:sched_switch//pull=9900/vm=cc7dc866-e8a6-49c8-9231-1bfba97e8849/,\
  kvm:kvm_exit//untraced/ --cycle --order -p 177789 --tsc --tsc-offset 0xffa01f12ce0cc8aa \
  -i 1000 --than 1s --detail=sametid

On the Guest, sample the sched:sched_switch event, first convert it to tsc timestamp,
and then push it to the /dev/virtio-ports/org.qemu.perf0 device.

On the Host, pull events from port 9900, which is connected to the Guest's org.qemu.perf0
device. The vm= attr will replace the cpu/pid/tid of the event and obtain the instance
corresponding to multi-trace.

For multi-trace, with the -p 177789 option, events are attached to 177789 all threads.
From this, the process of obtaining an instance:

  cpu => Host tid => ins
      |           ` host tid to instance. Use -p 177789 option.
      ` Guest event cpu to Host tid. Use vm= attr, Guest cpu and Host tid are one-to-one.

Signed-off-by: duanery <corcpp@foxmail.com>