prof_dev_clone(), clone to get a new prof_dev
The new prof_dev will use the profiler and environment of the parent. And attach
to the new cpu_map and thread_map. Others such as: timer, order, convert, can all
be rebuilt using env. Profiler-specific memory, it is up to the profiler to decide
how to share it with the new prof_dev. And, forwarded to the same target as the
parent device.

The cloned prof_dev is independent of the parent.

prof_dev_is_cloned(), used in profiler to determine whether prof_dev is cloned,
return parent prof_dev. Used to share Profiler-specific memory with the parent.

prof_dev_at_top(), determine whether the cloned device is at the top, used in
shared Profiler-specific memory scenarios. The top device can do interval output.

prof_dev_top_cloned(), get the top clone device. Used in forwarding scenarios to
identify the root forwarding source device.

1) The cloned device may be closed first. Need to support calling prof_dev_close()
in enabled(), sample(), which requires prof_dev to support reference counting.

2) After the cloned device is closed, its events may be cached in the order buffer
of the forwarding target, and they may be returned to the cloned device. This also
requires prof_dev to support reference counting to ensure that it is not released.
Get the reference when queuing to the order cache, and put the reference after
flushing.

So, prof_dev_get(), prof_dev_put(), are used for this.

prof_dev_flush() not only flushes the ringbuffer, but also flushes events in the
order buffer.

As a principle, if the forwarding target is closed, the forwarding source must be
closed. This includes cloned forwarding sources. In prof_dev_close(), the source
device must be closed. However, the child device created by tp_list_new() may also
be closed as a forwarding source. Therefore, uniformly close the forwarding source
and non-clone child devices in prof_dev_close(), not in tp_list_free().

Signed-off-by: duanery <corcpp@foxmail.com>