event-spread: The cpu for guest events may be [---].
When the Guest vcpu is bound to multiple Host cpu, the real host cpu of the Guest event
cannot be accurately obtained. Therefore, it will be assigned a value of -1 and [---]
will be displayed to avoid misleading.

[root@kvm ~]# virsh vcpupin $uuid
VCPU: CPU Affinity
----------------------------------
   0: 0-7
   1: 0-7
   2: 0-7
   3: 0-7
   4: 0-7
   5: 0-7
   6: 0-7
   7: 0-7

[root@kvm ~]# perf-prof trace -e sched:sched_switch//pull=9900/vm=$uuid/
Connected to 127.0.0.1:9900
2023-11-09 12:27:25.206865 G        CPU 0/KVM  80523 d... [---] 64555.365042: sched:sched_switch:

vm= attr is used to convert the Guest event field into the corresponding value of the Host.
Since vcpu is bound to 0-7, the cpu output is [---].

Signed-off-by: duanery <corcpp@foxmail.com>