multi-trace: add event_comparable(), samecpu supports cpu tracking.
event_comparable() determines whether `curr' can be compared with event1 and event2.
Events may come from Guest, Host-kernel, Host-forward. Among them, Guest-Guest and
Host-Host events can be compared with each other. However, when the Guest event
specifies vm ATTR, it can also be compared with the Host event.

Only events that can be compared can work with samecpu, samepid, and sametid. This
works even if the Guest event does not specify vm ATTR. event2 may be NULL, so it
cannot be compared with event2.

Currently, samecpu has three types of CPU-related events that can be printed.

- semecpu-event1. Events on the same CPU as event1.
- samecpu-event2. Events on the same CPU as event2.
- samecpu-track.  Events traversed by cpu tracking.

cpu tracking is to track the cpu where the pid of event1 is running on, and print the
events of this cpu. For general events, the CPU samples the event in the task's context,
therefore, the subsequent events of this CPU are printed. But there are some exceptions.

- sched:sched_wakeup, its target_cpu field specifies the running cpu after the task is
  awakened. Therefore, events of this cpu need to be tracked.

- sched:sched_migrate_task, its dest_cpu field specifies the running cpu after the task
  is migrated. Events for this cpu also need to be tracked.

- sched:sched_switch, tracks the execution and waiting of this task. Track events on thi
  CPU when executing, and stop tracking when waiting.

tp_target_cpu() is used for cpu tracking. Currently, task-state forwarding source events
are also supported.

The cpu of event1 and event2 may be [---], so only the samecpu-track can be used.

Signed-off-by: duanery <corcpp@foxmail.com>