multi-trace: Correctly handle out-of-order events.
When --order is enabled and --order-mem is set too small, events will be out of order.
Multi-trace will have strange output.

For two(A,B), where B is out of order, B will be added to the middle of the `timeline'
instead of the end. Within delay_two(), events after B on the timeline will be output.

For A->B->C, and enable the --cycle scenario. The order B, C, A(out of order) seen by
multi-trace will cause two (C, A) to be incorrect. In multi_trace_tryto_call_two(),
make sure C<A.

For A->B->C scenario. The order B, A(out of order), C seen by multi-trace will lose
two(B, C), because B will be replaced by A. In multi_trace_tryto_backup(), B is
replaced only when B<A.

Signed-off-by: duanery <corcpp@foxmail.com>