lib/perf: only link active perf_mmap
When evsel->own_cpus != NULL, and not equal to evlist->cpus. Only the intersection
of the two will be used in mmap_per_evsel(), and the corresponding perf_mmap, which
will cause part of evlist->mmap to be inactive.

If perf_mmap with idx=0 is inactive, evlist->mmap_first will not be set, causing
`perf_evlist__for_each_mmap' traversal to fail.

Signed-off-by: duanery <corcpp@foxmail.com>