rundelay: Track the running process and time between e1->e2.
For rundelay, getting the running time of each process on the same CPU helps to
understand why there is a delay. Using the sched:sched_switch event, on the same
CPU, the delay between two sched_switches is the process running time.

tp_oncpu() only obtains the current `pid' and `prev_comm' from the sched_switch
event. For the same CPU, when the pid changes, the running time of prev_comm is
calculated. The sampling timestamp of each sched_switch needs to be recorded for
such calculations.

Signed-off-by: duanery <corcpp@foxmail.com>