Clone the new prof_dev under the control of ptrace.
For prof_dev, after it is attached to a process (-p <pid>), there is no way to
track the newly forked child process. However, we can use ptrace to control the
forked child process(tracee), which is initially in a STOP state, and can notify
it to CONTINUE at the appropriate time.

The main idea is:
   When in the STOP state, clone a new prof_dev for the tracee.
   Then, notify it to CONTINUE.

   When tracee exits, close its prof_dev.

Initially, ptrace is used to trace the fork, vfork, and clone syscalls.
And re-inject all delivered signals.

To avoid stopping the process, we use PTRACE_SEIZE to attach to the process
and making it a tracee of the perf-prof(tracer). When it calls fork, tracer will
be notified. At the same time, it is in STOP initially. The tracer will first
continue its parent process, and then notify it to continue after obtaining the
new prof_dev using prof_dev_clone().

When tracee exits, it will notify the tracer and close its prof_dev. We do not
use PTRACE_O_TRACEEXIT, which can save 1 stop.

For Signal-delivery-stop, all signals delivered to tracee are first passed to
tracer and then fully injected into it. For SIGSTOP, the tracee will be stopped
until the SIGCONT signal is received.

Signed-off-by: duanery <corcpp@foxmail.com>