Detach from the tracer when prof_dev is closed.
When prof_dev is closed in various scenarios, clean up all the devs it clones,
and detach from the tracer(perf-prof). If prof_dev has a workload, kill it.
Otherwise, detach.

Whether we kill it or detach, we must stop the tracee first. Send a
PTRACE_INTERRUPT request, and after it stops, continue to send a PTRACE_KILL or
PTRACE_DETACH request. And then dereference prof_dev (@dev_users, @refcount).
Finally, prof_dev and its clones will be freed.

Each thread can only be attached to one tracer, which means that it can only be
grabbed by one prof_dev. Therefore, the prof_dev::ptrace_list variable is
introduced to link the ptrace threads(struct pid_node) for this dev.

Signed-off-by: duanery <corcpp@foxmail.com>