Safer detach from the tracer when prof_dev is closed.
The CLD_TRAPPED event returned through waitpid/waitid is completely random. During
detach, it is extremely likely that the fork's child process will still be a tracee.

Like this:
    PTRACE 17645 FORK 17646
            CONT 17645 signo 0
    PTRACE 17645 FORK 17647
            CONT 17645 signo 0
    PTRACE 17645 FORK 17648
            CONT 17645 signo 0
    PTRACE_INTERRUPT 17645     # detach 17645 and 17646 17647 17648
    PTRACE 17645 PTRACE_INTERRUPT | SIGCONT | initial new
            DETACH 17645 signo 0
    PTRACE 17646 PTRACE_INTERRUPT | SIGCONT | initial new
            DETACH 17646 signo 0
    ...

And like this:
    PTRACE 17645 FORK 17646
            CONT 17645 signo 0
    CHILD 17645 EXITED return 0
    PTRACE 17646 PTRACE_INTERRUPT | SIGCONT | initial new
            CONT 17646 signo 0

For the FORK stop, we need to make it completely independent, referencing or copying
everything needed for the child. This includes:

    parent pid_node
        The parent node, and its prof_dev. Prevent it from being released
        by subsequent DETACH and EXITED. After cloning the child's prof_dev,
        the parent node is no longer needed.

    parent detach/workload
        Child nodes will dereference the parent node in time, and at least
        the 'workload' field needs to be copied.

For the parent node, it is necessary to link all the child nodes in the initial
ptrace-stop state and notify its detach during ptrace_detach().

For the workload, when sending a PTRACE_INTERRUPT request, the workload is stopped.
The PTRACE_KILL sent next cannot inject the SIGKILL signal into the workload. This
is a bug, kill the workload directly instead.

Fixes: 03aebf6 ("Detach from the tracer when prof_dev is closed.")

Signed-off-by: duanery <corcpp@foxmail.com>