ptrace: Save one PTRACE_INTERRUPT for parent.
When the parent clones the child process, it is in the ptrace-stop state and can do
DETACH and CONT. No matter which one, its child will reference the parent node.

For a parent node, it might be:

    PTRACE_INTERRUPT 97252
    PTRACE 97252 CLONE 97257
            DETACH 97252         # put
or:
    PTRACE 97252 CLONE 97257
            CONT 97252
    CHILD 97252 EXITED return 0  # put
or:
    PTRACE 97252 CLONE 97257
            CONT 97252
    PTRACE_INTERRUPT 97252
            DETACH 97252         # put

Whether it is DETACH or CONT, it will only be dereferenced once. After the child node
enters the ptrace-stop state, it will dereference the parent node. Finally, the parent
node reference drops to 0 and is released.

Signed-off-by: duanery <corcpp@foxmail.com>