Simplify: When EXITED, ptraced prof_dev cannot be closed.
For ptraced prof_dev, after ptrace_exited() returns, is either closed or is still in
use. If it is closed, it will be removed from `prof_dev_list'. If it is in use, its
`ptrace_list' is not empty. Therefore, simplify.

Simplify: 7c04d24 ("When EXITED, prof_dev controlled by ptrace cannot be closed directly.")
Signed-off-by: duanery <corcpp@foxmail.com>