ptrace: Multiple prof_dev can attach to the same pid.
Multiple prof_dev are attached to one pid, which means that for the initial ptrace-stop
child, each prof_dev must be cloned. But each prof_dev can be closed independently, and
detach is performed after the last prof_dev is closed.

When a pid EXITED, all prof_dev attached to it need to be closed.

Signed-off-by: duanery <corcpp@foxmail.com>