order: heap sort: initial version.
Deprecated ordered-events implementation.

Events on each ringbuffer are ordered, so heap sort is the optimal solution.

Take the first event of all ringbuffers to build a min heap. After popping the top
event, continue to take the next event from the corresponding ringbuffer, heap it,
and continue popping until all ringbuffers becomes empty.

When a ringbuffer becomes empty, it may generate events again during the sorting
process. So each time it can only be sorted to the latest event of the ringbuffer.

According to the causal relationship, when I see the latest event, events that
occurred before the latest on other ringbuffers must have been written. In
implementation, we can only view the latest events of each ringbuffer one by one
and sort them. There is no way to sort events after latest.

For events after latest, they can only be put back into the ringbuffer. And wait
for other ringbuffers to be sorted.

Signed-off-by: duanery <corcpp@foxmail.com>