order: heap sort: add pauses to identify the lost period.
Why do this?

Since all ringbuffer events are sorted together, these events are a whole. When
a ringbuffer loses events, other ringbuffer events will also be unsafe when they
are within this lost period.

                        lost
    perf_mmap A: - -A-|=======|-A- -    # Have a lost
    perf_mmap B: - -B-|- -B- -|-B- -    # No lost
                      |   |   `lost_end
                      |   ` B is unsafe to perf_mmap A
                      `lost_start

We must identify `lost_start' and `lost_end' and notify the profiler in advance,
so that it can correctly handle events in this lost period.

However, the kernel will pre-output the lost event only when a new event is
generated. This means that the later the new event is, the later the lost event
is. No new events are generated, and the lost event will never be seen. We can
hardly see lost_start and lost_end events at the same time.

When heap sorting, we can only predict the possible location of the lost event,
and keep the `lost_start' event on the ringbuffer, and wait for the generation of
new events. New events continue to occur, the lost events need to be continuously
predicted.

In the kernel, when we see new events, we can always see lost events.

Predict, use wakeup_watermark as safe space. The kernel writes exceed the watermark
and perf-prof is woken up to handle the event. With more writing, once the safe
space is not enough, it is considered that there may be a lost event. Then start
predict.

If a possible loss is predicted, pause and stop the heap sort to wait for new events.

This way, we can guarantee that the lost event will only be discovered during the
order_process(). And can see the lost_start and lost_end events in advance, and can
continue to sort the events of other ringbuffers during the lost period.

Signed-off-by: duanery <corcpp@foxmail.com>